requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.22

# Async database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Task queue
celery==5.4.0
redis==5.0.4

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# LLM clients
anthropic==0.26.0
openai==1.30.1
google-generativeai==0.5.4

# Input parsing
openpyxl==3.1.2

# Config
pyyaml==6.0.1

# Utilities
aiofiles==23.2.1
httpx==0.27.0
python-slugify==8.0.1
python-frontmatter==1.1.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
aiosqlite==0.19.0
